Using `fps` to replace `FrameRate` parameter.
`FrameRate` parameter is succeeded from AVS2-P2 reference software `RD`, and it means *FrameRateCode* in the bit-stream. This parameter usually cause error results in the first glance, since we would treat it as number of frames in one second.
The new parameter, `fps` would be utilized in later version, and `FrameRate` would be deprecated.